requirements.txt
# Requirements for developing and running the minigame PPO experiments
stable-baselines3
sb3-contrib
gymnasium
torch
pysc2
protobuf>=3.20,<4.0
absl-py
opencv-python
tensorboard
matplotlib
numpy
